unified alfa 1 with 1.0.1-a1

+++ app/Http/Middleware/Admin.php
@@ -18,7 +18,7 @@ class Admin
     public function handle(Request $request, Closure $next){
         $menuName = 'investigate';
         $title = '';
-        if(Auth::user()->is_admin !== 1){
+        if( (int)Auth::user()->is_admin !== 1 ){
             return response()->view('platform.admin-check', compact('title', 'menuName'));
         }
         return $next($request);
